requirements.txt
onnxruntime
numpy
scipy
torchvision
pillow
pytest
